+++ objc-sys/src/exception.rs
@@ -2,39 +2,36 @@
 //! Apple: `objc-exception.h`
 //! GNUStep: `eh_personality.c`, which is a bit brittle to rely on, but I
 //!   think it's fine...
-#[cfg(any(not(objfw), feature = "unstable-exception"))]
 use core::ffi::c_void;
-#[cfg(apple_new)]
+#[cfg(any(doc, apple_new))]
 use std::os::raw::c_int;
 #[cfg(feature = "unstable-exception")]
 use std::os::raw::c_uchar;
 
-#[cfg(apple_new)]
+#[cfg(any(doc, apple_new))]
 use crate::objc_class;
 use crate::objc_object;
 
 /// Remember that this is non-null!
-#[cfg(apple_new)]
+#[cfg(any(doc, apple_new))]
 pub type objc_exception_matcher =
     unsafe extern "C" fn(catch_type: *mut objc_class, exception: *mut objc_object) -> c_int;
 
 /// Remember that this is non-null!
-#[cfg(apple_new)]
+#[cfg(any(doc, apple_new))]
 pub type objc_exception_preprocessor =
     unsafe extern "C" fn(exception: *mut objc_object) -> *mut objc_object;
 
 /// Remember that this is non-null!
-#[cfg(apple_new)]
+#[cfg(any(doc, apple_new))]
 pub type objc_uncaught_exception_handler = unsafe extern "C" fn(exception: *mut objc_object);
 
 #[cfg(objfw)]
 pub type objc_uncaught_exception_handler =
     Option<unsafe extern "C" fn(exception: *mut objc_object)>;
 
-/// Only available on macOS.
-///
 /// Remember that this is non-null!
-#[cfg(all(apple_new, target_os = "macos"))]
+#[cfg(any(doc, all(apple_new, target_os = "macos")))]
 pub type objc_exception_handler =
     unsafe extern "C" fn(unused: *mut objc_object, context: *mut c_void);
 
@@ -48,6 +45,7 @@ extern_c_unwind! {
     ///
     /// [`objc-exception.h`]: https://github.com/apple-oss-distributions/objc4/blob/objc4-818.2/runtime/objc-exception.h
     pub fn objc_exception_throw(exception: *mut objc_object) -> !;
+
     #[cfg(apple_new)]
     pub fn objc_exception_rethrow() -> !;
 
@@ -56,15 +54,15 @@ extern_c_unwind! {
 }
 
 extern_c! {
-    #[cfg(any(gnustep, apple_new))]
+    #[cfg(any(doc, gnustep, apple_new))]
     pub fn objc_begin_catch(exc_buf: *mut c_void) -> *mut objc_object;
-    #[cfg(any(gnustep, apple_new))]
+    #[cfg(any(doc, gnustep, apple_new))]
     pub fn objc_end_catch();
 
-    #[cfg(apple_old)]
+    #[cfg(any(doc, apple_old))]
     pub fn objc_exception_try_enter(exception_data: *const c_void);
 
-    #[cfg(apple_old)]
+    #[cfg(any(doc, apple_old))]
     pub fn objc_exception_try_exit(exception_data: *const c_void);
 
     // objc_exception_extract
@@ -72,28 +70,26 @@ extern_c! {
     // objc_exception_get_functions
     // objc_exception_set_functions
 
-    #[cfg(apple_new)]
+    #[cfg(any(doc, apple_new))]
     pub fn objc_setExceptionMatcher(f: objc_exception_matcher) -> objc_exception_matcher;
-    #[cfg(apple_new)]
+    #[cfg(any(doc, apple_new))]
     pub fn objc_setExceptionPreprocessor(
         f: objc_exception_preprocessor,
     ) -> objc_exception_preprocessor;
-    #[cfg(any(apple_new, objfw))]
+    #[cfg(any(doc, apple_new, objfw))]
     pub fn objc_setUncaughtExceptionHandler(
         f: objc_uncaught_exception_handler,
     ) -> objc_uncaught_exception_handler;
 
-    /// Only available on macOS.
-    #[cfg(all(apple_new, target_os = "macos"))]
+    #[cfg(any(doc, all(apple_new, target_os = "macos")))]
     pub fn objc_addExceptionHandler(f: objc_exception_handler, context: *mut c_void) -> usize;
-    /// Only available on macOS.
-    #[cfg(all(apple_new, target_os = "macos"))]
+    #[cfg(any(doc, all(apple_new, target_os = "macos")))]
     pub fn objc_removeExceptionHandler(token: usize);
 
     // Only available when ENABLE_OBJCXX is set, and a useable C++ runtime is
     // present when building libobjc2.
     //
-    // #[cfg(gnustep)]
+    // #[cfg(any(doc, gnustep))]
     // pub fn objc_set_apple_compatible_objcxx_exceptions(newValue: c_int) -> c_int;
 
     /// Call the given function inside an Objective-C `@try/@catch` block.